feat: implement profiler with new framework APIs
